classification by routers update
